Fixed some bugs Group-structure

+++ JSON/src/main/java/org/it_academy/MK_JD2_90_22/json/services/GroupService.java
@@ -7,18 +7,14 @@ import org.it_academy.MK_JD2_90_22.json.dao.entity.Group;
 import org.it_academy.MK_JD2_90_22.json.dto.group.GroupName;
 import org.it_academy.MK_JD2_90_22.json.dto.group.GroupRefresh;
 import org.it_academy.MK_JD2_90_22.json.exceptions.service.GroupIllegalNameException;
-import org.it_academy.MK_JD2_90_22.json.exceptions.service.GroupNullPointerException;
-import org.it_academy.MK_JD2_90_22.json.exceptions.service.StudentsIllegalIdException;
 import org.it_academy.MK_JD2_90_22.json.services.api.ICRUDGroupService;
 
 import java.util.List;
-import java.util.Objects;
 
 public class GroupService implements ICRUDGroupService {
 
     private static final GroupService instance;
     private static final IValidationDao validator = ValidationDao.getInstance();
-
     private static final GroupDao dao = GroupDao.getInstance();
 
     static {
@@ -30,10 +26,13 @@ public class GroupService implements ICRUDGroupService {
 
     @Override
     public void save(GroupName groupName) {
-        isNull(groupName);
+
+        if (groupName.getName().isEmpty()) {
+            throw new GroupIllegalNameException("Не введено название группы для сохранения");
+        }
 
         if (validator.isExistGroup(groupName.getName())) {
-            new GroupIllegalNameException();
+            throw new GroupIllegalNameException();
         }
 
         dao.save(groupName);
@@ -41,7 +40,14 @@ public class GroupService implements ICRUDGroupService {
 
     @Override
     public void delete(GroupName groupName) {
-        isNull(groupName);
+
+        if (groupName.getName().isEmpty()) {
+            throw new GroupIllegalNameException("Не введено название группы для удаления");
+        }
+
+        if (!validator.isExistGroup(groupName.getName())) {
+            throw new GroupIllegalNameException("Такой группы не существует");
+        }
 
         dao.delete(groupName);
     }
@@ -53,8 +59,9 @@ public class GroupService implements ICRUDGroupService {
 
     @Override
     public Group get(long id) {
+
         if (id <= 0) {
-            throw new StudentsIllegalIdException();
+            throw new GroupIllegalNameException("ID < 1");
         }
 
         return dao.get(id);
@@ -62,16 +69,23 @@ public class GroupService implements ICRUDGroupService {
 
     @Override
     public void update(GroupRefresh groupRefresh) {
-        isNull(groupRefresh);
+
+        String oldName = groupRefresh.getOldName();
+        String newName = groupRefresh.getNewName();
+
+        if (oldName == null || oldName.isEmpty()) {
+            throw new GroupIllegalNameException("Не введено старое название группы");
+        }
+
+        if (newName == null || newName.isEmpty()) {
+            throw new GroupIllegalNameException("Не введено новое название группы");
+        }
+
+        if (!validator.isExistGroup(oldName)) {
+            throw new GroupIllegalNameException("Такой группы не существует");
+        }
 
         dao.update(groupRefresh);
-    }
-
-    private void isNull(Object o) {
-
-        if (Objects.isNull(o)){
-            new GroupNullPointerException();
-        }
     }
 
     public static GroupService getInstance() {
